LPS-60722 - merge BaseTranslationManagerTag and TranslationManagerTag
