+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-taglib/src/main/java/com/liferay/dynamic/data/mapping/taglib/servlet/taglib/TranslationManagerTag.java
@@ -14,7 +14,11 @@
 
 package com.liferay.dynamic.data.mapping.taglib.servlet.taglib;
 
-import com.liferay.dynamic.data.mapping.taglib.servlet.taglib.base.BaseTranslationManagerTag;
+import com.liferay.dynamic.data.mapping.taglib.servlet.ServletContextUtil;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.jsp.JspException;
+import javax.servlet.jsp.PageContext;
 
 /**
  * @author Eduardo Lundgren
@@ -22,5 +26,119 @@ import com.liferay.dynamic.data.mapping.taglib.servlet.taglib.base.BaseTranslati
  * @author Nathan Cavanaugh
  * @author Julio Camarero
  */
-public class TranslationManagerTag extends BaseTranslationManagerTag {
+public class TranslationManagerTag extends com.liferay.taglib.util.IncludeTag {
+
+	@Override
+	public int doStartTag() throws JspException {
+		setAttributeNamespace(_ATTRIBUTE_NAMESPACE);
+
+		return super.doStartTag();
+	}
+
+	public java.util.Locale[] getAvailableLocales() {
+		return _availableLocales;
+	}
+
+	public java.lang.String getDefaultLanguageId() {
+		return _defaultLanguageId;
+	}
+
+	public java.lang.String getEditingLanguageId() {
+		return _editingLanguageId;
+	}
+
+	public java.lang.String getId() {
+		return _id;
+	}
+
+	public boolean getInitialize() {
+		return _initialize;
+	}
+
+	public boolean getReadOnly() {
+		return _readOnly;
+	}
+
+	public void setAvailableLocales(java.util.Locale[] availableLocales) {
+		_availableLocales = availableLocales;
+
+		setScopedAttribute("availableLocales", availableLocales);
+	}
+
+	public void setDefaultLanguageId(java.lang.String defaultLanguageId) {
+		_defaultLanguageId = defaultLanguageId;
+
+		setScopedAttribute("defaultLanguageId", defaultLanguageId);
+	}
+
+	public void setEditingLanguageId(java.lang.String editingLanguageId) {
+		_editingLanguageId = editingLanguageId;
+
+		setScopedAttribute("editingLanguageId", editingLanguageId);
+	}
+
+	public void setId(java.lang.String id) {
+		_id = id;
+
+		setScopedAttribute("id", id);
+	}
+
+	public void setInitialize(boolean initialize) {
+		_initialize = initialize;
+
+		setScopedAttribute("initialize", initialize);
+	}
+
+	@Override
+	public void setPageContext(PageContext pageContext) {
+		super.setPageContext(pageContext);
+
+		setServletContext(ServletContextUtil.getServletContext());
+	}
+
+	public void setReadOnly(boolean readOnly) {
+		_readOnly = readOnly;
+
+		setScopedAttribute("readOnly", readOnly);
+	}
+
+	@Override
+	protected void cleanUp() {
+		super.cleanUp();
+
+		_availableLocales = null;
+		_defaultLanguageId = null;
+		_editingLanguageId = null;
+		_id = null;
+		_initialize = true;
+		_readOnly = false;
+	}
+
+	@Override
+	protected String getPage() {
+		return _PAGE;
+	}
+
+	@Override
+	protected void setAttributes(HttpServletRequest request) {
+		setNamespacedAttribute(request, "availableLocales", _availableLocales);
+		setNamespacedAttribute(request, "defaultLanguageId", _defaultLanguageId);
+		setNamespacedAttribute(request, "editingLanguageId", _editingLanguageId);
+		setNamespacedAttribute(request, "id", _id);
+		setNamespacedAttribute(request, "initialize", _initialize);
+		setNamespacedAttribute(request, "readOnly", _readOnly);
+	}
+
+	protected static final String _ATTRIBUTE_NAMESPACE = "liferay-ddm:translation-manager:";
+
+	private static final String _PAGE =
+		"/translation_manager/page.jsp";
+
+	private java.util.Locale[] _availableLocales = null;
+	private java.lang.String _defaultLanguageId = null;
+	private java.lang.String _editingLanguageId = null;
+	private java.lang.String _id = null;
+	private boolean _initialize = true;
+	private boolean _readOnly = false;
+
 }
